requirements.txt
flask
redis
prometheus_client
opentelemetry-api
opentelemetry-sdk
werkzeug
